requirements.txt
openai>=1.15
tiktoken>=0.4.0
langchain-openai>=0.1
langchain-community>=0.1
langchain-chroma
chromadb>=0.5
tree_sitter==0.20.2
tree_sitter_languages>=1.6
gitpython>=3.1
fastapi
uvicorn[standard]
python-dotenv
